+++ modules/hanna-css/src/css/StatusTag.css.ts
@@ -4,7 +4,7 @@ import { buildVariables } from '../lib/cssutils.js';
 import { hannaVars as vars } from '../lib/hannavars.js';
 
 export const StatusTagVariables = buildVariables(
-  ['background', 'borderColor', 'lightColor', 'height', 'paddingInline'],
+  ['background', 'borderColor', 'lightColor'],
   'StatusTag'
 );
 const Stv = StatusTagVariables.vars;
@@ -13,61 +13,55 @@ export default css`
   @media screen {
     .StatusTag {
       ${StatusTagVariables.declare({
-        background: vars.color_faxafloi_25,
-        borderColor: vars.color_faxafloi_50,
-        lightColor: vars.color_faxafloi_100,
-        height: vars.space_3,
-        paddingInline: vars.space_1,
+        borderColor: vars.color_suld_50,
+        background: vars.color_suld_25,
+        lightColor: vars.color_suld_100,
       })};
       display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      gap: 6px;
-      box-sizing: border-box;
-      height: ${Stv.height};
-      padding: 0 ${Stv.paddingInline};
-      border-radius: ${vars.space_0$5};
-      border: 1px solid ${Stv.borderColor};
-      background: ${Stv.background};
-      color: ${vars.color_suld_200};
       font: ${vars.font_label};
       font-weight: ${vars.font_weight__bold};
+      align-items: center;
+      padding: ${vars.space_0$5} ${vars.space_1};
+      border: 1px solid ${Stv.borderColor};
+      border-radius: ${vars.space_0$5};
+      background: ${Stv.background};
+      color: ${vars.color_suld_200};
       vertical-align: middle;
       white-space: nowrap;
 
-      margin-bottom: ${vars.space_2};
       margin-right: ${vars.space_2};
+      margin-bottom: ${vars.space_2};
     }
 
     .StatusTag--large {
-      ${StatusTagVariables.override({
-        height: vars.space_4,
-        paddingInline: vars.space_1$5,
-      })};
+      padding: ${vars.space_1} ${vars.space_1$5};
     }
 
-    .StatusTag__light {
+    .StatusTag::before {
+      content: '';
       color: ${Stv.lightColor};
       background: currentColor;
       border-radius: 50%;
       flex: 0 0 auto;
       width: ${vars.space_1};
       height: ${vars.space_1};
-      display: block;
+      margin-right: 6px; /* Figma design has 6px space */
+    }
+    .StatusTag--light--off::before {
+      content: none;
     }
 
     .StatusTag__label {
       display: inline-block;
     }
 
-    .StatusTag--color--grey {
+    .StatusTag--color--blue {
       ${StatusTagVariables.override({
-        borderColor: vars.color_suld_50,
-        background: vars.color_suld_25,
-        lightColor: vars.color_suld_100,
+        background: vars.color_faxafloi_25,
+        borderColor: vars.color_faxafloi_50,
+        lightColor: vars.color_faxafloi_100,
       })};
     }
-
     .StatusTag--color--green {
       ${StatusTagVariables.override({
         borderColor: vars.color_ellidaardalur_100,
@@ -75,7 +69,6 @@ export default css`
         lightColor: vars.color_ellidaardalur_100,
       })};
     }
-
     .StatusTag--color--yellow {
       ${StatusTagVariables.override({
         borderColor: vars.color_nautholsvik_100,
@@ -83,7 +76,6 @@ export default css`
         lightColor: vars.color_nautholsvik_100,
       })};
     }
-
     .StatusTag--color--red {
       ${StatusTagVariables.override({
         borderColor: vars.color_heidmork_100,