+++ modules/hanna-react/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+
+## 0.10.171
+
+_2026-04-02_
+
 - feat: Add component `StatusTag`
 - feat: Add component `Timeline`
 - `utils`: